Unbreak definition execution. Previously things like
(define (foo) "foo") (define bar (foo)) bar

were broken.

The problem stemmed from ProcCall objects on the trampoline that looked like this:

(proc0 _1 ...)

where proc0 was bound in environment X, _1 was bound in environment Y, and
neither X nor Y was a descendant of the other. Since the operator is evaluated
in the same environment as its operands, this was bound to fail.

My "solution" is to start setting Environment objects on Continuation objects
in addition to Continuable objects. The hope is that we can remove Environment
setting on Continuables.