Implement left-recursive procedure calls correctly, as in
((lambda (x)
  ((lambda (y)
    ((lambda (z)
      (+ x y z z z)) 3)) 2)) 1)

Previously, we were evaluating the operator _during_ desugaring,
which was causing all kinds of scope resolution problems. So now
we merely desugar the operator, like so:

((f x) (g y) (h z)) => (g y [_0 (h z [_1 (f x [_2 (_2 _0 _1 [_3 ...])])])])

Regardless of this win, the tree of Environments during trampolining is
becoming a nightmare. We are resorting to kludges to set the enclosing
environments correctly.