Fix subtle bug in desugaring
(define foo (lambda () ...))

which was implemented correctly, versus

(define (foo) ...)

which wasn't. The defined procedure was being bound to the formal parameter
at desugar time. What should happen is it should be bound to a new name at
desugar time, then bound to the formal parameter during execution of the
definition on the trampoline.

I'm surprised this bug didn't cause more chaos.