Major architectural change: no eager "CPSification" of procedure calls.
Previously, we were CPSifying everything at desugar time that parsed as
a procedure call. For example,

(foo (+ x y))

is "CPSified" as

(+ x y [_0 (foo _0 ...)])

At evaluation time, on the trampoline, if it turns out that foo is actually
a macro, we needed somewhat involved logic to reconstruct the datum

(+ x y)

that should be passed to foo's pattern matching facility. Plus, the
effort that went into CPSifying what turned out not to be a procedure call
is wasted.

In the new architecture, (foo (+ x y)) desugars merely as

(foo (+ x y) [_0 ...])

At evaluation time, if it turns out foo is a procedure, that's when we
CPSify its arguments.

This change had the positive effect of turning hundreds of lines of
bad programming into dead code. However, it also slows down the interpreter
(likely due to having to desugar the same datum over and over on the
trampoline --- this should be amenable to caching), and it causes a unit
test covering recursive procedure calls to fail. Whatever the solutions to
these problems, I doubt they'll be as obscure as the code I deleted to get
here.