Reorganized how we construct the standard environments at bootstrap time.
We can now successfully load syntax definitions like let (though this
requires an actual web server for the ajax request).

This commit exposes the fact that we still haven't gotten the semantics of
macro definition environments right. For example, evaluation of

(let ((x 1)) (+ x x))

fails because the environment in which let is defined is further down the
stack than the environment in which + is defined. We need some kind of
post-bootstrap squash procedure that will boil all the environments down to
one flat one.