Simplify implementation of top-level bindings in preparation for REPL stuff.
Example REPL session:

1> (define x 1)
2> (+ x x)

Each line is handled by a different call to the trampoline. For this to
work, we have to make sure line 2 can see the binding introduced by line 1.
Previously, we had the trampoline return an environment that would be
considered the new "top level" environment, which would be used by subequent
calls. I found it simpler just to send all top-level bindings straight
to the top of the environment chain. Then line 2 doesn't need to know where
line 1 left off.