Rewrite of the environment-assigning system to hopefully be clearer.
I realized that a major source of complexity in the environment-assigning
system was having all these Continuable objects which all had assignable
Environments.

The system should not be that complex. If the current environment is X,
there are really only three cases to consider:

1) You're at a primitive procedure call. The environment stays X.
2) You're at a non-primitive procedure call. Make a new environment Y
   pointing back to X, set the procedure body's environment to Y,
   but remember after the procedure body is done to go back to X.
3) You're at a macro use. Make a new environment Z pointing back to X,
   transcribe the macro, set the transcription's environment to Z,
   but remember after the transcript is done executing to go back to X.

In particular, branches should be oblivious to environments.

At this point, the major rewrite is done and all the tests are working
(except a legacy failure I haven't looked at for weeks). There are still
a couple hacks and the whole thing needs to be better documented.