First draft of supporting quasiquotations as arguments.
Committed a unit test that shows what I need to work on next:
desugaring a quasiquotation may or may not result in further CPSification.

Example: with (car `(1 2)), after sequenceOperands we should have merely

(id (1 2) [_0 (car _0 [_1 ...])])

But with (car `(1 ,(+ 2 3))) we should have something like

(+ 2 3 [_0 (id (1 _0) [_1 (car _1 [_2 ...])])])

We don't know what will happen until we perform the CPSification, but in
the current design (LocalStructure constructor), we have to choose ahead of
time.

Also got rid of a tree traversal in resetDesugars(). I'm not sure
why it doesn't cause any test failures...