First draft of renaming identifiers in macro transcriptions to avoid capture.
I'm not sure my heuristic for what to rename is correct, but it seems to
work for letrec (which is the only macro that requires renaming so far).

There should be an easy performance win in this commit by renaming
the identifiers prior to transcription, rather than transcribing first and
then walking the tree to rename everything.