Support (set! var expr) on the trampoline (as opposed to lambda emulation).
Previously we were transforming

(set! var expr) ...

to something like

((lambda (var) ...) expr)

except that we had special logic to ensure var was already bound.

We used to transform all sorts of stuff into lambdas internally, but
today I replaced _all_ of that with more appropriate abstractions. (See
following commits.) Once they were out of the way, it was easy to make
a special case for set! on the trampoline.