Implement top-level definitions with set!, not let*.
Implementing top-level definitions with let*, while a big improvement
over my previous ad-hoc logic, was complicated because it created hundreds
of bindings and required me to partition the program into definitions and
expressions, and run the two separately. (!!!) It also caused problems with
setting the definition environments of macros correctly. So on the whole,
this is a big win.

There is still one clear bug: because syntax definitions are still installed
at desugar time, they will always be shadowed by variable definitions,
even if they occur later in the program text. This should not be hard
to fix.