Defensively copy patterns when creating SchemeMacros.
We have refrained from copying patterns when creating SchemeMacros for
months (at least since commit 608206d2), but recent changes to the parser
(commit e8ca699c) showed that if the text of a macro definition was reparsed,
that could change the SchemeMacro defined from it.

We probably need to do an audit of where Datum.prototype.clone() is being
used, and unfortunately more importantly, where it needs to be used but is
not currently.