Preliminary support for rebinding fundamental identifiers (define, if, ...).
Since fundamental identifiers can be rebound, it really does not make sense
to have a "parser" that assigns nonterminal labels to nodes in a parse tree.
Getting rid of the parser (while at the same time keeping all the
well-formedness constraints it gives us) is more work than I can do right now,
so for the time being I've just written a hack that renames suspicious
identifiers and re-parses. It should incur little overhead in the common case
when these identifiers are not rebound.