Support simple uses of force and delay.
(There's nothing special semantically about them, it just seems
they exposed a bug in my closure-setting logic, specifically with
doing closure lookups in a non-recursive way. I've patched this for
the moment by copying explicitly up to a least common ancestor, but this
seems very expensive and may not be correct either. A rewrite of the closure
stuff may be in order.)