Make relationships among bindings clearer.
Previously, we had all sorts of hacks to get the right bindings visible at
the right time. This commit takes a big step towards getting rid of the
hacks and getting environment-setting to work like you'd imagine. Example:

(((lambda (x)
	(lambda (y)
		(+ x y))) 10) 11)

At a high level, the lambda-x creates a new environment, binds x = 11,
and passes it to the lambda-y. The lambda-y creates a new environment
on top of that, binds y = 10, and passes it to (+ x y).

There are some nuances in order to support multiple active procedure calls,
but that is the main idea.

TODO: I had to temporarily disable tail recursion to get this to work, but
it shouldn't be too hard to reenable. The switch (in tryNonPrimitiveProcedure)
is clear.