Never unwrap string or character Datums when storing in Environments.
This is because string, character, and identifier Datums all would unwrap
to JavaScript strings, and it would be impossible to reconstruct which
was which. I chose to unwrap identifier Datums alone; hopefully they are the
common case.

This also required changing the implementations of the primitive
procedures that take strings or chars as arguments (basically had to
add in a manual unwrap for each such argument).