Replace custom scanner with regexes.
I would have thought the performance improvement would be dramatic.
Even if it doesn't pan out, though, it's still a big win in terms
of simplicity.

TODO: the scanner currently ignores what should be scan errors. For example:

hello ] world

The ] is happily skipped over as if it were not present. Shouldn't be hard
to detect and throw.