mockterm: let the browser reflow text, don't do it ourselves.
MockTerminal's attempt at reflowing text around numColumns was sometimes
off by a word (perhaps due to rounding error in the calculation of character
width). When this happened it was worse than no reflowing at all because
a word would end up on a line by itself. So, best the leave the browser to its
own devices. We can fiddle with the CSS white-space if need be.