java: add extern to fix node repl.
CompilationUnits.NODE_REPL was failing at runtime. This was because
r5js.platform.node.Terminal called readline.Interface#on, which is
actually defined in a superclass, events.EventEmitter, which wasn't
being pulled in by the externs. The call to `on` was renamed by the
compiler and therefore failed at runtime.