eval: Pipeline: experiment with impl-hiding.
This commit puts the Pipeline interface and implementation into
the same ES6 module, exporting only the interface and a factory method.
This has the advantage of hiding the implementation, but the cost
is that the module that provides the interface requires all the
dependencies for the implementation (because they are the same module).
This makes it impossible to use the interface to break circular deps.

I'd like to use this pattern as much as possible, falling back to
@package implementations and public factory methods when necessary.