runtime: rm IEnvironment#isImplementedBy, etc.
@interfaces don't exist at runtime (they are removed by the compiler),
so any code that explicitly wants to check if a value implements an
interface has to use some other API. that was the original intent of
IEnvironment#addImplementation() and isImplementedBy().

however, Environment is the only implementation of IEnvironment.
(the interface was extracted to break a dep, not to allow multiple
impls.) and all the code that was calling IEnvironment#isImplementedBy()
was in places where it is possible to write `instanceof Environment`
directly.