convert //js:error to ts_library.
https://github.com/bazelbuild/rules_typescript/pull/112 is an
in-progress PR that will add a closure_ts_library or closure_ts_binary
rule to rules_typescript. but the review hasn't progressed for a month.

in
https://github.com/Ubehebe/rules_typescript/commit/dd4169d3e1d6b8f6eda1d7c1f29fb9368542f16a,
i patched in the PR and got it to work. it's not pretty, but it allows
per-target conversion of closure_js_library to closure_ts_library (and
eventually just ts_library).

closure_ts_library transpiles ts to es6. the main question for this
codebase is how es6 modules interoperate with goog.modules. jscompiler
has some amount of interop support: see
https://github.com/google/closure-compiler/wiki/JS-Modules and
https://github.com/google/closure-compiler/issues/2644. in particular,
es6 modules (`import`) can be required from
goog.modules (`goog.require`) by going through a third module system,
commonjs (`require`). the module path that works for jscompiler and the
bazel rules is ugly (e.g.
`require('/js/error_collect_es6_sources.es6/node_modules/__main__/js/error')`),
but it is enough to migrate the codebase to typescript.