ast: put SimpleDatum and Identifier back in their own modules.
in b9629c, i discovered that partial migration of a class hierarchy to
ts caused two different copies of the base class to be included in the
compiled output, leading to runtime nonsense. the workaround in that
commit was to move the subclass being migrated into the same module as
the base class. i followed this strategy with two Datum subclasses in
05168f and 6f58d2.

but i have discovered a better workaround. the scope that seems
relevant isn't a module but a ts_library target. that is, the subclasses
can be in separate .ts source files (and therefore modules) as long as
they appear together as srcs in the same ts_library target.

this inflicts less damage on the commit history when migrating a js file
to ts, so i prefer it.