[ts] migrate r5js.ast.CompoundDatum.
items of interest:

the same failure mode as b9629c: no compilation error but strange test
failure. the failure was in the syntax-rebinding-tests in
other-tests.scm:

```
(define if +)
(if 1 2 3)
```

should eval to 6, but it was evaling to 2. this implicated
CompoundDatum.fixParserSensitiveIds. debugging revealed the same issue
as b9629c: two copies of the Nonterminal class were present at runtime,
causing the === comparison in Datum.hasParse() to fail.

the workaround in b9629c was to move the duplicated class into the same
ts_library. this is not an option in the current case, since there are
targets that depend on Nonterminal but not Datum.

i believe the root cause of both of these issues is that
closure_ts_library violates the one version rule. if you build
//js/ast:datum_collect_es6_sources, you will see that it generates the
transitive closure of its dependencies under its own path, e.g.

```
bazel-bin/js/ast/datum_collect_es6_sources.es6/node_modules/__main__/js/parse/nonterminals.js
```

when other closure_js_library targets depend on //js/parse:nonterminals,
that is a different copy than the one //js/ast:datum generates.

the workaround in this commit is to delete the
closure_ts_library (//js/parse:nonterminals) and have closure_js_library
dependents `require()` the copy from the genfiles tree created by
//js/ast:datum. this is not ideal, since the closure_js_library targets
now have an implicit dep on //js/ast:datum. but it should be a temporary
state, and it gets us further along the path of eliminating the
closure_*_library targets.