[ts] migrate r5js.TokenStream.
items of interest: same failure mode as ce5eab and b9629c.
in this case, a second copy of the AST classes was showing up in
//js/scan:scan_collect_es6_sources, causing instanceof checks to fail.

the workaround in this case is a little harder, since the source file
that caused the failure is in a different package (//js/scan
vs. //js/ast). i could move token_stream.ts to //js/ast, but that would
pollute the revision history. instead, this commit sets up a genrule in
//js/ast that copies token_stream.ts into the package. i'm surprised
that it works, since token_stream.ts contains relative import
paths, but it does.