BUILD: add a bunch of closure compiler suppressions.
i'm not sure how this happened. bazel is supposed to be hermetic, using
a pinned version of closure compiler. without source or deps changes,
the build shouldn't suddenly break.

the most likely culprit is the fact that i've been developing against
local forks of deps for a while (local_repositories instead of
git_repositories). the switch to typescript is so imminent that i don't
care to investigate more.