rm //js/ast:ugh genrule hack.
this is the first of several commits unwinding the genrules i created to
prevent the duplicate prototype runtime failures during the js->ts
migration. the desired end state is a source tree that can be built
either incrementally with bazel or traditionally with any tool that
interprets the top-level tsconfig.json as the root of ts project.