s/declare type/export type/g.
type declarations are part of a library's api (equivalent to closure
externs). i was just using them to get around some codegen issues when
the ts sources had to coexist with js sources.

now that the types are exported instead of ambient, they have to be
imported, so this commit adds dozens of import statements. which is good
and explicit.