macro: use Sets for templateRenameCandidate data structures.
the data structure at head, {[key: string]: boolean}, is idiomatic in js
and probably performs better. but Sets express intent better and
hopefully will become idiomatic.

TODO: find a way for tsc to treat bracket access on a Set or Map as a
compile error. migrating an object to a Set or Map is error-prone.