update infrastructure. no code changes.
- bump rules_nodejs 0.8.0 => 0.27.8
- bump rules_typescript 0.12.3 => 0.27.8
- add an explicit package.json and yarn.lock
- replace ts_web_test targets with jasmine_node_test targets (for now)

there has been a significant change in the recommended workflow for
rules_typescript. instead of being a first-class bazel repository,
rules_typescript is now declared as an npm dependency in
package.json. it is usable as a bazel repo under the name
@npm_bazel_typescript through the magic of
install_bazel_dependencies(), invoked in this repo's WORKSPACE.