WORKSPACE: add mockterm external repo. not used yet.
i am working at recreating the Makefile (deleted in 50fae44) as a
bunch of bazel genrules/etc. the ultimate aim is to have a top-level
tar_pkg target whose output (a .tar.gz file) can be uploaded to any
static server. much like the Makefile used to work, but in a different
world.