rules+test: rewrite scheme_source as a starlark rule.
at head, scheme_source is a macro that expands to a genrule and a
ts_library. the genrule embeds the .scm source inside a string literal
in a .ts file, and the ts_library uses that .ts file as its src.

this is overcomplicated. any ts_library that needs to use the
generated .ts files can list them as srcs. and once we don't need to
generate ts_library targets, the macro can be rewritten more clearly
as a rule.