ui: get web repl mostly working.
MockTerminal maintains a stack of "interpreters". when interpretation
is called for, it interrogates the stack of interpreters until one of
them gives an answer.

at head, this api was poorly typed; the MockTerminal directly invoked
elements of the interpreter stack as though they were js
functions. this commit introduces an Interpreter typedef that makes it
a little clearer what the contract is.