ui: remove fancy rotary nav.
the rotary nav looked cool, but did not have a strong reason for
existing. it was a nonstandard ui and very much desktop-only.

now that i have removed most of the content from the index page, it
has even less reason for existing.

this commit also removes a few helper classes (TextResizer,
VisibilityManager) that interacted with the rotary nav in some
way. their existence seems like a code smell to me now; it should be
possible to handle most ui issues with modern css.