ui: rm mobile site.
the mobile site was based on jquery mobile, a project that seems to
have become an evolutionary dead end (see
https://github.com/jquery/jquery-mobile/issues/8612). i also suspect
that browser support has advanced to the degree that it should be
possible to serve an attractive mobile ui from the same html+css as
the desktop site.

(what the mobile ux should be is a fundamental question, since a repl
is fundamentally not appropriate as a mobile ui.)

removing the mobile stuff also paves the way for getting rid of the
xsl pipeline, another bit of antiquated technology.