ui: rm xsl pipeline.
now that the mobile website is gone, there's no reason to generate the
(single) website from a source xml file. just keep it in version
control as html.

xsltproc and friends are interesting technically, but have proved to
be a bit of an evolutionary dead end. but the main reason for this
commit is just that i don't need the processing anymore.

TODO: a build-time html linter would still be useful.