rules: scheme_source: embed inside es6 template literals.
part of the interpreter's boot process is loading and evaluating the
.scm sources defining the standard syntax and procedures. this is
implemented by embedding the .scm sources in a javascript string
literal, then including the string literals in the main interpreter
compilation unit.

scheme_source is the starlark rule that does this. at head, it
consists of a bunch of sed expressions that munge the .scm sources
into a form appropriate for embedding into a (pre-es6) js string
literal. this includes replacing newlines with \n (since pre-es6
string literals cannot be multiline), escaping quotes, etc.

this commit switches to targeting es6 template literals. these are
delimited by ` (backticks) and allow multiline text. the sed
expressions are consequently simplified. (scheme does use backticks
for quasiquotations, so these still have to be escaped.)