switch from commonmark to pandoc to generate site.
the documentation for this project should contain a running
interpreter. ideally, the interpreter could be embedded in the
README.md somehow, so that it would render automatically in the github
ui, making a separate site superfluous.

unfortunately, the opportunities for interactivity on a
github-rendered markdown file are quite limited. it strips script and
iframe tags; the best i can do on the github ui is probably an <img>
tag showing a rendering of the interpreter. so a separate site,
derived from the README.md but also embedding the interpreter, is
desirable.

this commit is prefactoring. it ditches commonmark in favor of pandoc,
which allows more flexibility in injecting stuff into rendered
output (through the use of custom templates). pandoc already has bazel
rules that appear to work out of the box. hooray!